+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add `iosxr_fpd` resource and data source
 - Add `iosxr_extcommunity_cost_set` resource and data source
 - Add `iosxr_rd_set` resource and data source
+- Add `contact` and `location` attributes to `iosxr_snmp_server` resource and data source
 
 ## 0.2.1